# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "大脑训练网页开发",
        "title": "API 文档",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/memory/matrix": {
            "get": {
                "description": "生成一组需要记忆的方格位置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "瞬间记忆"
                ],
                "summary": "生成记忆训练矩阵",
                "parameters": [
                    {
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ],
                        "type": "string",
                        "default": "easy",
                        "description": "难度级别",
                        "name": "difficulty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routers.MemoryMatrixResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/memory/scores": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "保存用户记忆训练成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "瞬间记忆"
                ],
                "summary": "提交记忆训练成绩",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour_token\u003e",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "成绩数据",
                        "name": "score",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routers.MemoryScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/schulte/matrix": {
            "get": {
                "description": "为舒尔特表格提供随机数字矩阵",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "舒尔特矩阵"
                ],
                "summary": "舒尔特表格",
                "parameters": [
                    {
                        "maximum": 10,
                        "minimum": 3,
                        "type": "integer",
                        "default": 5,
                        "description": "矩阵大小（如 5 表示 5x5 矩阵）",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回矩阵数据",
                        "schema": {
                            "$ref": "#/definitions/routers.MatrixResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/schulte/scores": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "保存用户舒尔特表格游戏成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "成绩管理"
                ],
                "summary": "提交成绩",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour_token\u003e",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "成绩数据",
                        "name": "score",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routers.SchulteScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routers.MatrixResponse": {
            "type": "object",
            "properties": {
                "matrix": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "routers.MemoryMatrixResponse": {
            "type": "object",
            "properties": {
                "answerTime": {
                    "description": "返回 回答时间",
                    "type": "number"
                },
                "difficulty": {
                    "description": "返回 等级难度",
                    "type": "string"
                },
                "displayTime": {
                    "description": "返回 展示时间",
                    "type": "number"
                },
                "positions": {
                    "description": "返回的要亮的 15 个数字",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "totalCells": {
                    "description": "返回一共所有的单元格",
                    "type": "integer"
                }
            }
        },
        "routers.MemoryScoreRequest": {
            "type": "object",
            "required": [
                "accuracy",
                "level",
                "successNum",
                "trainingNum",
                "userId"
            ],
            "properties": {
                "accuracy": {
                    "type": "number"
                },
                "level": {
                    "type": "string"
                },
                "successNum": {
                    "type": "integer"
                },
                "trainingNum": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "routers.SchulteScoreRequest": {
            "type": "object",
            "required": [
                "isPassed",
                "trainingNum",
                "userId"
            ],
            "properties": {
                "isPassed": {
                    "type": "boolean"
                },
                "trainingNum": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}